+++ CHANGELOG.md
@@ -9,6 +9,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- auto-save prompt content between page reloads
+- show thumbnails in chat history for uploaded fles
 
 ### Changed